step with retry
